Playing around with HeartBuds
Trying to get a clean signal, no high frequencies.

The heartbuds device is very crackly and the crackles are quite loud, so we need to filter out high and loud sounds. This is my attempt to do so. But, the lowpass filter is crackling, my attempts to remove loud noises are only working at the resolution of one callback by zeroing loud data blocks, the limiter is not responding quickly enough or the sample I am working from is buggy.

Maybe other folks know more than me about this.